src/python/requirements.txt
# Azure Communication Services - Python SDK Dependencies
# ==============================================================================
# Install with: pip install -r requirements.txt
# ==============================================================================

# Core Azure SDK packages
azure-identity>=1.15.0
azure-core>=1.30.0

# Azure Communication Services SDKs
azure-communication-identity>=1.4.0
azure-communication-sms>=1.1.0
azure-communication-chat>=1.2.0
azure-communication-email>=1.0.0
azure-communication-phonenumbers>=1.2.0
azure-communication-callautomation>=1.1.0

# Azure Key Vault for secrets
azure-keyvault-secrets>=4.7.0

# Web framework for API
flask>=3.0.0
flask-cors>=4.0.0

# Async support
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Logging and monitoring
opencensus-ext-azure>=1.1.0
applicationinsights>=0.11.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
responses>=0.24.0

# Development tools
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0
pylint>=3.0.0
